--- bin/recon-mutil
@@ -1,13 +1,0 @@
-#!/bin/bash
-
-. `dirname $0`/../libexec/env.sh
-
-split_cli $@
-
-export MALLOC_ARENA_MAX=1
-
-java ${JAVA_OPTS-} -Xmx10536m -Xms6024m -XX:+UseSerialGC ${jvm_options[@]} \
-    -cp ${COATJAVA_CLASSPATH:-''} \
-    org.jlab.clas.reco.EngineMultiProcessor \
-    ${class_options[@]}
-
